Moved encoders into sub-directory, added server spinup
