+++ encoders/builtin/enc_test.go
@@ -1,15 +1,19 @@
-// Copyright 2012-2014 Apcera Inc. All rights reserved.
+// Copyright 2012-2015 Apcera Inc. All rights reserved.
 
-package nats
+package builtin_test
 
 import (
 	"bytes"
 	"testing"
 	"time"
+
+	"github.com/nats-io/nats"
+	"github.com/nats-io/nats/encoders/builtin"
+	"github.com/nats-io/nats/test"
 )
 
-func NewEConn(t *testing.T) *EncodedConn {
-	ec, err := NewEncodedConn(newConnection(t), DEFAULT)
+func NewEConn(t *testing.T) *nats.EncodedConn {
+	ec, err := nats.NewEncodedConn(test.NewDefaultConnection(t), nats.DEFAULT_ENCODER)
 	if err != nil {
 		t.Fatalf("Failed to create an encoded connection: %v\n", err)
 	}
@@ -17,17 +21,20 @@ func NewEConn(t *testing.T) *EncodedConn {
 }
 
 func TestConstructorErrs(t *testing.T) {
-	c := newConnection(t)
-	_, err := NewEncodedConn(nil, "default")
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
+	c := test.NewDefaultConnection(t)
+	_, err := nats.NewEncodedConn(nil, "default")
 	if err == nil {
 		t.Fatal("Expected err for nil connection")
 	}
-	_, err = NewEncodedConn(c, "foo22")
+	_, err = nats.NewEncodedConn(c, "foo22")
 	if err == nil {
 		t.Fatal("Expected err for bad encoder")
 	}
 	c.Close()
-	_, err = NewEncodedConn(c, "default")
+	_, err = nats.NewEncodedConn(c, "default")
 	if err == nil {
 		t.Fatal("Expected err for closed connection")
 	}
@@ -35,6 +42,9 @@ func TestConstructorErrs(t *testing.T) {
 }
 
 func TestMarshalString(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
@@ -48,7 +58,7 @@ func TestMarshalString(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_string", testString)
-	if e := wait(ch); e != nil {
+	if e := test.Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -57,6 +67,9 @@ func TestMarshalString(t *testing.T) {
 }
 
 func TestMarshalBytes(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
@@ -70,7 +83,7 @@ func TestMarshalBytes(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_bytes", testBytes)
-	if e := wait(ch); e != nil {
+	if e := test.Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -79,6 +92,9 @@ func TestMarshalBytes(t *testing.T) {
 }
 
 func TestMarshalInt(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
@@ -92,7 +108,7 @@ func TestMarshalInt(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_int", testN)
-	if e := wait(ch); e != nil {
+	if e := test.Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -101,6 +117,9 @@ func TestMarshalInt(t *testing.T) {
 }
 
 func TestMarshalInt32(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
@@ -114,7 +133,7 @@ func TestMarshalInt32(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_int", testN)
-	if e := wait(ch); e != nil {
+	if e := test.Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -123,6 +142,9 @@ func TestMarshalInt32(t *testing.T) {
 }
 
 func TestMarshalInt64(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
@@ -136,7 +158,7 @@ func TestMarshalInt64(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_int", testN)
-	if e := wait(ch); e != nil {
+	if e := test.Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -145,6 +167,9 @@ func TestMarshalInt64(t *testing.T) {
 }
 
 func TestMarshalFloat32(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
@@ -158,7 +183,7 @@ func TestMarshalFloat32(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_float", testN)
-	if e := wait(ch); e != nil {
+	if e := test.Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -167,6 +192,9 @@ func TestMarshalFloat32(t *testing.T) {
 }
 
 func TestMarshalFloat64(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
@@ -180,7 +208,7 @@ func TestMarshalFloat64(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_float", testN)
-	if e := wait(ch); e != nil {
+	if e := test.Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -189,6 +217,9 @@ func TestMarshalFloat64(t *testing.T) {
 }
 
 func TestMarshalBool(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
@@ -200,7 +231,7 @@ func TestMarshalBool(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_bool", false)
-	if e := wait(ch); e != nil {
+	if e := test.Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -209,6 +240,9 @@ func TestMarshalBool(t *testing.T) {
 }
 
 func TestExtendedSubscribeCB(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 
@@ -227,7 +261,7 @@ func TestExtendedSubscribeCB(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish(subject, testString)
-	if e := wait(ch); e != nil {
+	if e := test.Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -236,6 +270,9 @@ func TestExtendedSubscribeCB(t *testing.T) {
 }
 
 func TestExtendedSubscribeCB2(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 
@@ -258,7 +295,7 @@ func TestExtendedSubscribeCB2(t *testing.T) {
 		ch <- true
 	})
 	ec.PublishRequest(oSubj, oReply, testString)
-	if e := wait(ch); e != nil {
+	if e := test.Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -267,6 +304,9 @@ func TestExtendedSubscribeCB2(t *testing.T) {
 }
 
 func TestRawMsgSubscribeCB(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 
@@ -276,7 +316,7 @@ func TestRawMsgSubscribeCB(t *testing.T) {
 	oSubj := "cb_args"
 	oReply := "foobar"
 
-	ec.Subscribe(oSubj, func(m *Msg) {
+	ec.Subscribe(oSubj, func(m *nats.Msg) {
 		s := string(m.Data)
 		if s != testString {
 			t.Fatalf("Received test string of '%s', wanted '%s'\n", s, testString)
@@ -290,7 +330,7 @@ func TestRawMsgSubscribeCB(t *testing.T) {
 		ch <- true
 	})
 	ec.PublishRequest(oSubj, oReply, testString)
-	if e := wait(ch); e != nil {
+	if e := test.Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -299,6 +339,9 @@ func TestRawMsgSubscribeCB(t *testing.T) {
 }
 
 func TestEncRequest(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 
@@ -315,6 +358,9 @@ func TestEncRequest(t *testing.T) {
 }
 
 func TestEncRequestReceivesMsg(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 
@@ -322,7 +368,7 @@ func TestEncRequestReceivesMsg(t *testing.T) {
 		ec.Publish(reply, "I can help!")
 	})
 
-	var resp Msg
+	var resp nats.Msg
 
 	err := ec.Request("help", "help me", &resp, 500*time.Millisecond)
 	if err != nil {
@@ -331,6 +377,9 @@ func TestEncRequestReceivesMsg(t *testing.T) {
 }
 
 func TestAsyncMarshalErr(t *testing.T) {
+	s := test.RunDefaultServer()
+	defer s.Shutdown()
+
 	ec := NewEConn(t)
 	defer ec.Close()
 
@@ -343,18 +392,18 @@ func TestAsyncMarshalErr(t *testing.T) {
 		// This will never get called.
 	})
 
-	ec.Conn.Opts.AsyncErrorCB = func(c *Conn, s *Subscription, err error) {
+	ec.Conn.Opts.AsyncErrorCB = func(c *nats.Conn, s *nats.Subscription, err error) {
 		ch <- true
 	}
 
 	ec.Publish(subject, testString)
-	if e := wait(ch); e != nil {
+	if e := test.Wait(ch); e != nil {
 		t.Fatalf("Did not receive the message: %s", e)
 	}
 }
 
 func TestEncodeNil(t *testing.T) {
-	de := &DefaultEncoder{}
+	de := &builtin.DefaultEncoder{}
 	_, err := de.Encode("foo", nil)
 	if err != nil {
 		t.Fatalf("Expected no error encoding nil: %v", err)
@@ -362,7 +411,7 @@ func TestEncodeNil(t *testing.T) {
 }
 
 func TestDecodeDefault(t *testing.T) {
-	de := &DefaultEncoder{}
+	de := &builtin.DefaultEncoder{}
 	b, err := de.Encode("foo", 22)
 	if err != nil {
 		t.Fatalf("Expected no error encoding number: %v", err)
