+++ encoders/builtin/gob_enc.go
@@ -1,6 +1,6 @@
-// Copyright 2013 Apcera Inc. All rights reserved.
+// Copyright 2013-2015 Apcera Inc. All rights reserved.
 
-package nats
+package builtin
 
 import (
 	"bytes"
